try to register new customer on back with errors

+++ backend/src/main/java/com/netcracker/edu/backend/service/CustomerEntityService.java
@@ -8,5 +8,5 @@ public interface CustomerEntityService {
     CustomerEntity saveCustomer(CustomerEntity customer);
     Optional<CustomerEntity> getCustomerById(long id);
     Iterable<CustomerEntity> getAllCustomers();
-    void deleteCustomer(Integer id);
+    void deleteCustomer(long id);
 }
